Only show untagged posts when selected
In order to avoid adding a lot of logic for searching through posts we
simply hide the render of the post component until an untagged post is
found. This feels like a terrible way of accomplishing this, but either
something exceptionally clever can be found, or it will require putting
the logic here in the app or in the backend. Either way it incurs the
same amount of Pinboard API requests.